fix(windows): check version for backdrop support
